Move the maneuver registry into the combat module
module/actor/maneuver.js defines the GURPS maneuver list, but a maneuver
is a thing you take on your turn in combat, not a property of an actor --
the file sits in module/actor/ only because the maneuver arrives on an
actor as an ActiveEffect.

With the maneuver settings now registered in module/combat/, leaving the
registry behind meant module/actor/ and module/combat/ imported each
other to answer one question about the On Target source book. Moving it
makes that an ordinary intra-module import.
