+++ lib/moustachewax.js
@@ -2,7 +2,7 @@
 
 import { extractOtfs } from '../module/otf/extract-otfs.js'
 import { MoveModes } from '../module/actor/actor.js'
-import Maneuvers from '../module/actor/maneuver.js'
+import Maneuvers from '../module/combat/maneuver.js'
 import * as HitLocations from '../module/hitlocation/hitlocation.js'
 import { TokenActions } from '../module/token-actions.js'
 import { multiplyDice } from '../module/utilities/damage-utils.js'
