+++ module/chat/chat-processors.js
@@ -6,7 +6,7 @@ import { parselink } from '../../lib/parselink.js'
 import { escapeHtml, isNiceDiceEnabled, makeRegexPatternFrom, splitArgs, wait } from '../../lib/utilities.js'
 import { ChatProcessors } from '../../module/chat.js'
 import { ActorImporter } from '../actor/actor-importer.js'
-import Maneuvers from '../actor/maneuver.js'
+import Maneuvers from '../combat/maneuver.js'
 import { AnimChatProcessor } from '../chat/anim.js'
 import SlamChatProcessor from '../chat/slam.js'
 import StatusChatProcessor from '../chat/status.js'
