+++ module/gurps.js
@@ -63,7 +63,7 @@ import { initialize_i18nHelper, translate } from '../lib/i18n.js'
 import { parseDecimalNumber } from '../lib/parse-decimal-number/parse-decimal-number.js'
 import { GGADebugger } from '../utils/debugger.js'
 import { EffectModifierControl } from './actor/effect-modifier-control.js'
-import Maneuvers from './actor/maneuver.js'
+import Maneuvers from './combat/maneuver.js'
 import { AddMultipleImportButton } from './actor/multiple-import-app.js'
 import GurpsActiveEffectConfig from './effects/active-effect-config.js'
 import GurpsActiveEffect from './effects/active-effect.js'
